pyproject.toml
[project]
name = "proven"
version = "0.1.0"
description = "LLM-driven stepwise refinement with Dafny verification"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Lionel Di Giacomo"},
]
readme = "README.md"
keywords = ["formal-verification", "dafny", "llm", "correct-by-construction", "stepwise-refinement"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/melek/proven"
Repository = "https://github.com/melek/proven"

[project.scripts]
proven = "proven.__main__:main"
